Add files via upload
Rules for couplet links.

KEY
Directionality:
Part after underscore indicates directionality. For example, "_AB" implies that an answer in couplet A implies an answer in couplet B, but not vice-versa. "_BA" implies the opposite, and "_R" implies that the link is forcing both ways. Couplet pairs with no underscore are reserved for "EQ" and "XOR".

AND: Statement 1 implies statement 2
XORP: Statement 1 implies NOT statement 2
XORN: NOT statement 1 implies statement 2
NOR: NOT statement 1 implies NOT statement 2.

Examples:
XORP_AB: A implies !B (exclusively)
NOR_R: !A implies !B, and !B implies !A
XORN_BA: !B implies A
EQ: A is always equal to B. (A<->B; !A<->!B)
XOR: A is always the negation of B. (A<->!B, !A<->B)

These rules may be used to generate a comprehensive dataset to evaluate LLM performance in natural language to conversion within this project.